move start_watching method to SpecWorker
